Uisng Context, Bug fixed, and styling Tab Nav
The bug was nothing major

Styled the Tab nav a bit
I noticed that createBottomTabNavigator and createMaterialTopTabNavigator has thesame prop to pass to the Tab.Navigation called tabBarOptions.
More on that here
https://reactnavigation.org/docs/material-top-tab-navigator#tabbaroptions
https://reactnavigation.org/docs/bottom-tab-navigator#tabbaroptions

Implemented React Context
This helps us pass the decks state to the children components from the App.js easily!